Peak型最適化とデータベーススキーマ移行対応
- Peak型をnumpy.float64に変更（12.37%の性能改善を確認）
- 性能分析スクリプト（peak_performance_analysis.py）を追加
- データベーススキーマ自動移行機能を実装
- 既存データベースのmetaカラム不足問題を解決
- migrate_database.pyユーティリティを追加

matsu582のGitHubコメント対応完了

+++ src/audio_fingerprinter.py
@@ -39,9 +39,9 @@ except ImportError:
 @dataclass
 class Peak:
     """時間-周波数領域のスペクトルピークを表現"""
-    time: float
-    frequency: float
-    amplitude: float
+    time: np.float64
+    frequency: np.float64
+    amplitude: np.float64
 
 
 @njit(cache=True, parallel=False)  # parallel=Falseに変更（順序保証のため）
@@ -232,9 +232,9 @@ class SpectrogramAnalyzer:
                 
                 if magnitude[f_idx, t_idx] == np.max(neighborhood):
                     peak = Peak(
-                        time=float(times[t_idx]),
-                        frequency=float(frequencies[f_idx]),
-                        amplitude=float(magnitude[f_idx, t_idx])
+                        time=np.float64(times[t_idx]),
+                        frequency=np.float64(frequencies[f_idx]),
+                        amplitude=np.float64(magnitude[f_idx, t_idx])
                     )
                     peaks.append(peak)
         
@@ -266,9 +266,9 @@ class SpectrogramAnalyzer:
             # Peak オブジェクト生成（元の実装と同じ）
             peaks = [
                 Peak(
-                    time=float(times[t_idx]), 
-                    frequency=float(frequencies[f_idx]), 
-                    amplitude=float(magnitude[f_idx, t_idx])
+                    time=np.float64(times[t_idx]), 
+                    frequency=np.float64(frequencies[f_idx]), 
+                    amplitude=np.float64(magnitude[f_idx, t_idx])
                 )
                 for f_idx, t_idx in zip(peak_f_indices, peak_t_indices)
             ]
@@ -297,7 +297,7 @@ class SpectrogramAnalyzer:
                 if f_start <= f_idx < f_end and t_start <= t_idx < t_end:
                     valid.append((f_idx, t_idx))
             # Peakオブジェクトを作成
-            peaks = [Peak(time=float(times[t]), frequency=float(frequencies[f]), amplitude=float(magnitude[f, t]))
+            peaks = [Peak(time=np.float64(times[t]), frequency=np.float64(frequencies[f]), amplitude=np.float64(magnitude[f, t]))
                      for f, t in valid]
             if debug:
                 logger.info(f"Candidate points: {candidates_found}")
@@ -424,7 +424,7 @@ class HashGenerator:
         
         # ピーク密度フィルタリングを適用
         filtered_peaks = self._filter_peaks_by_density(peaks, debug)
-        sorted_peaks = sorted(filtered_peaks, key=lambda p: p.time)
+        sorted_peaks = sorted(filtered_peaks, key=lambda p: float(p.time))
         
         if debug:
             self._log_peak_info(sorted_peaks, logger)
@@ -644,7 +644,7 @@ class HashGenerator:
             logger.info(f"Peak count before density filtering: {len(peaks)}")
         
         # ピークを時間順にソート
-        sorted_peaks = sorted(peaks, key=lambda p: p.time)
+        sorted_peaks = sorted(peaks, key=lambda p: float(p.time))
         
         # 最小間隔に基づくフィルタリング
         filtered_peaks = []
